requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
beautifulsoup4
requests 